Reorganized code in packages
