+++ src/main/kotlin/de/qaware/cloud/nativ/k8s/cluster/ClusterDeployments.kt
@@ -1,16 +1,7 @@
-package de.qaware.cloud.nativ.k8s
+package de.qaware.cloud.nativ.k8s.cluster
 
-import io.fabric8.kubernetes.api.KubernetesHelper
 import io.fabric8.kubernetes.api.model.extensions.Deployment
-import io.fabric8.kubernetes.client.KubernetesClient
-import io.fabric8.kubernetes.client.KubernetesClientException
-import io.fabric8.kubernetes.client.Watcher
-import org.apache.deltaspike.core.api.config.ConfigProperty
-import org.slf4j.Logger
-import javax.annotation.PostConstruct
 import javax.enterprise.context.ApplicationScoped
-import javax.enterprise.event.Event
-import javax.inject.Inject
 
 /**
  * This class handles the Deployments.
