+++ src/main/kotlin/de/qaware/cloud/nativ/k8s/k8s/KubernetesPods.kt
@@ -21,8 +21,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package de.qaware.cloud.nativ.k8s
+package de.qaware.cloud.nativ.k8s.k8s
 
+import de.qaware.cloud.nativ.k8s.cluster.ClusterNode
 import io.fabric8.kubernetes.api.KubernetesHelper
 import io.fabric8.kubernetes.api.model.Pod
 import io.fabric8.kubernetes.client.KubernetesClient
