+++ src/main/kotlin/de/qaware/cloud/nativ/k8s/launchpad/MidiSystemProducer.kt
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package de.qaware.cloud.nativ.k8s
+package de.qaware.cloud.nativ.k8s.launchpad
 
 import javax.enterprise.context.ApplicationScoped
 import javax.enterprise.inject.Default
